+++ client/src/components/ProfileArtist/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable consistent-return */
 import React, { useState, useEffect } from 'react';
 import propTypes from 'prop-types';
 import Axios from 'axios';
@@ -8,6 +9,8 @@ import AddProduct from '../AddProduct';
 import AuthorizationContext from '../../Contexts/AuthorizationContext';
 import PaintingSection from '../PaintingSection/PaintingsSection';
 import EditArtistProfile from '../EditArtistProfile';
+import AddReview from '../ReviewForm';
+import ReviewContainer from '../ReviewForm/ReviewContainer';
 
 import './style.css';
 
@@ -19,6 +22,9 @@ function Profile({ match }) {
   const [error, setError] = useState(null);
   const [loaded, setLoaded] = useState(false);
   const [ArtistImg, setArtistImg] = useState();
+  const [reviewVisible, setReviewVisible] = useState(false);
+  const [reviews, setReviews] = useState();
+
   const cloudinaryLink =
     'https://res.cloudinary.com/dacf3uopo/image/upload/v1593353472/';
   const { artistId } = match.params;
@@ -38,6 +44,29 @@ function Profile({ match }) {
         e = 'تعذر جلب بيانات الفنان';
       }
       setError(e);
+    }
+  };
+
+  const getArtistReviews = async (id) => {
+    try {
+      const { data } = await Axios.get(`/api/v1/review/${id}`);
+      if (data.statusCode === 200) {
+        setReviews(data.data);
+      }
+    } catch (err) {
+      let e;
+      if (err.response) {
+        switch (err.response.data.message) {
+          case 'Artist ID should be number':
+            e = 'لا يمكن العثور على هذا الفنان';
+            break;
+          default:
+            e = 'فشلت العملية, يرجى المحاولة لاحقاً';
+        }
+      } else {
+        e = 'فشلت العملية, يرجى المحاولة لاحقاً';
+      }
+      message.error(e);
     }
   };
 
@@ -62,13 +91,19 @@ function Profile({ match }) {
     }
   };
 
+  const hideReview = () => setReviewVisible(false);
+
   useEffect(() => {
     getArtistProfile(artistId);
-  }, []);
+  }, [artistId]);
 
   useEffect(() => {
     getAllPainting(artistId);
-  }, []);
+  }, [artistId]);
+
+  useEffect(() => {
+    getArtistReviews(artistId);
+  }, [artistId]);
 
   const hideForm = () => setShowForm(false);
   const hideFormEdit = () => setShowFormEdit(false);
@@ -212,15 +247,20 @@ function Profile({ match }) {
           )}
         </div>
         <div className="left-header">
-          <AuthorizationContext.Consumer>
-            {({ user }) =>
-              isAuth(user) && (
-                <Button onClick={() => setShowFormEdit(true)}>
-                  تعديل بيانات الحساب
-                </Button>
-              )
-            }
-          </AuthorizationContext.Consumer>
+          <div>
+            {' '}
+            {reviews && (
+              <ReviewContainer className="reviewsSlider" reviews={reviews} />
+            )}
+          </div>
+
+          {reviewVisible && (
+            <AddReview
+              reviewVisible={reviewVisible}
+              hideReview={hideReview}
+              artistID={artistId}
+            />
+          )}
           {showFormEdit && (
             <EditArtistProfile
               profileData={profileData}
@@ -234,13 +274,27 @@ function Profile({ match }) {
       </header>
       <div className="painting-container">
         <AuthorizationContext.Consumer>
-          {({ user }) =>
-            isAuth(user) && (
-              <Button onClick={() => setShowForm(true)}>
-                اضافة لوحة جديدة
-              </Button>
-            )
-          }
+          {({ user }) => {
+            if (isAuth(user)) {
+              return (
+                <>
+                  <Button onClick={() => setShowForm(true)}>
+                    اضافة لوحة جديدة
+                  </Button>
+                  <Button onClick={() => setShowFormEdit(true)}>
+                    تعديل بيانات الحساب
+                  </Button>
+                </>
+              );
+            }
+            if (user.role === 'customer') {
+              return (
+                <Button onClick={() => setReviewVisible(true)}>
+                  أضف تقييم الفنان
+                </Button>
+              );
+            }
+          }}
         </AuthorizationContext.Consumer>
         {showForm && (
           <AddProduct
