+++ client/src/components/ReviewForm/style.css
@@ -1,5 +1,4 @@
 .reviewSlider {
   width: 300px;
   margin: 10% auto;
-  border: 1px solid black;
 }
