+++ server/database/data/build.sql
@@ -50,7 +50,7 @@ CREATE TABLE artist
         artist_id INTEGER REFERENCES artist(id) ON UPDATE CASCADE ON DELETE CASCADE,
         customer_id INTEGER REFERENCES customer(id) ON UPDATE CASCADE ON DELETE CASCADE,
         rate INTEGER NOT NULL,
-        details text NOT NULL,
+        details text,
         review_date TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP
 
     );
